// src/lib/theme/types.ts
export type ThemeMode = 'light' | 'dark';

export interface ThemeIdentity {
  siteName: string;
  markLetters?: string;
  markShape?: 'shield' | 'hex';
  mastheadEyebrow?: string;
  footerSectionTitle?: string;
  footerDisclaimer?: string;
}

export interface ThemeTokens {
  appBg: string;
  appBgDecor: string;
  appText: string;
  appTextMuted: string;
  mastheadBg: string;
  mastheadBorder: string;
  mastheadTitle: string;
  mastheadEyebrow: string;
  mastheadIconStart: string;
  mastheadIconEnd: string;
  navPillBg: string;
  navPillBorder: string;
  navPillText: string;
  navPillHoverBg: string;
  navPillHoverBorder: string;
  navPillHoverText: string;
  navDropdownBg: string;
  navDropdownBorder: string;
  navDropdownText: string;
  navDropdownHoverBg: string;
  navDropdownHoverText: string;
  utilityButtonBg: string;
  utilityButtonBorder: string;
  utilityButtonText: string;
  utilityButtonHoverBg: string;
  utilityButtonHoverBorder: string;
  utilityButtonHoverText: string;
  menuPanelBg: string;
  menuPanelBorder: string;
  menuPanelText: string;
  menuPanelMutedText: string;
  menuPanelHoverBg: string;
  pageHeaderBg: string;
  pageHeaderBorder: string;
  pageLabel: string;
  pageTitle: string;
  pageDescription: string;
  pageActionBg: string;
  pageActionBorder: string;
  pageActionText: string;
  pageActionHoverBg: string;
  pageActionHoverText: string;
  cardBg: string;
  cardBgAccent: string;
  cardBgDeep: string;
  cardText: string;
  cardLabel: string;
  cardTitle: string;
  articleCardBg: string;
  articleCardBorder: string;
  articleCardHoverBorder: string;
  articleCardTitle: string;
  articleCardText: string;
  articleCardMutedText: string;
  articleCardBadgeBg: string;
  articleCardBadgeText: string;
  articleCardCaptionBorder: string;
  articleCardPlaceholderBg: string;
  articleCardPlaceholderBorder: string;
  articleCardPlaceholderText: string;
  surfaceBorder: string;
  surfaceShadow: string;
  buttonPrimaryBg: string;
  buttonPrimaryText: string;
  buttonPrimaryHoverBg: string;
  buttonSecondaryBg: string;
  buttonSecondaryText: string;
  buttonSecondaryHoverBg: string;
  buttonNeutralBg: string;
  buttonNeutralText: string;
  buttonNeutralHoverBg: string;
  buttonDangerBg: string;
  buttonDangerText: string;
  buttonDangerHoverBg: string;
  inputBg: string;
  inputBorder: string;
  inputText: string;
  inputPlaceholder: string;
  inputFocusRing: string;
  formLabel: string;
  adminSurfaceBg: string;
  adminSurfaceBorder: string;
  adminSurfaceMutedBg: string;
  adminText: string;
  adminMutedText: string;
  adminLink: string;
  adminLinkHover: string;
  sidebarLinkEvenBg: string;
  sidebarLinkOddBg: string;
  sidebarLinkHoverBg: string;
  statusErrorBg: string;
  statusErrorBorder: string;
  statusErrorText: string;
  statusErrorIconBg: string;
  statusErrorIconText: string;
  statusWarningBg: string;
  statusWarningBorder: string;
  statusWarningText: string;
  statusWarningIconBg: string;
  statusWarningIconText: string;
  statusInfoBg: string;
  statusInfoBorder: string;
  statusInfoText: string;
  statusInfoIconBg: string;
  statusInfoIconText: string;
  statusSuccessBg: string;
  statusSuccessBorder: string;
  statusSuccessText: string;
  statusSuccessIconBg: string;
  statusSuccessIconText: string;
  brandPrimary: string;
  brandAccent: string;
  fontSans: string;
  fontSerif: string;
  radiusCard: string;
  radiusPill: string;
}

export interface TenantThemeManifest {
  tenantSlug: string;
  themeName: string;
  defaultMode: ThemeMode;
  supports: ThemeMode[];
  identity: ThemeIdentity;
  modes: Record<ThemeMode, ThemeTokens>;
}

export interface ResolvedTheme {
  manifest: TenantThemeManifest;
  mode: ThemeMode;
  identity: ThemeIdentity;
  tokens: ThemeTokens;
}